Inline createX functions into createInstance in HostConfig
We don't need the hoistable type to use `is` since that's not supported.
