+++ packages/react-dom-bindings/src/client/ReactDOMComponent.js
@@ -15,7 +15,6 @@ import {
 } from '../events/EventRegistry';
 
 import {canUseDOM} from 'shared/ExecutionEnvironment';
-import hasOwnProperty from 'shared/hasOwnProperty';
 import {checkHtmlStringCoercion} from 'shared/CheckStringCoercion';
 
 import {
@@ -59,7 +58,6 @@ import {
 } from './CSSPropertyOperations';
 import {HTML_NAMESPACE, getIntrinsicNamespace} from './DOMNamespaces';
 import {getPropertyInfo} from '../shared/DOMProperty';
-import {DOCUMENT_NODE} from './HTMLNodeType';
 import isCustomComponent from '../shared/isCustomComponent';
 import possibleStandardNames from '../shared/possibleStandardNames';
 import {validateProperties as validateARIAProperties} from '../shared/ReactDOMInvalidARIAHook';
@@ -67,7 +65,6 @@ import {validateProperties as validateInputProperties} from '../shared/ReactDOMN
 import {validateProperties as validateUnknownProperties} from '../shared/ReactDOMUnknownPropertyHook';
 
 import {
-  enableTrustedTypesIntegration,
   enableCustomElementPropertySupport,
   enableClientRenderFallbackOnTextMismatch,
   enableHostSingletons,
@@ -79,25 +76,8 @@ import {
 } from '../events/DOMPluginEventSystem';
 
 let didWarnInvalidHydration = false;
-let didWarnScriptTags = false;
-
-let warnedUnknownTags: {
-  [key: string]: boolean,
-};
 let canDiffStyleForHydrationWarning;
-
 if (__DEV__) {
-  warnedUnknownTags = {
-    // There are working polyfills for <dialog>. Let people use it.
-    dialog: true,
-    // Electron ships a custom <webview> tag to display external web content in
-    // an isolated frame and process.
-    // This tag is not present in non Electron environments such as JSDom which
-    // is often used for testing purposes.
-    // @see https://electronjs.org/docs/api/webview-tag
-    webview: true,
-  };
-
   // IE 11 parses & normalizes the style attribute as opposed to other
   // browsers. It adds spaces and sorts the properties in some
   // non-alphabetical order. Handling that would require sorting CSS
@@ -262,14 +242,6 @@ export function checkForUnmatchedText(
     // client rendering, up to the nearest Suspense boundary.
     throw new Error('Text content does not match server-rendered HTML.');
   }
-}
-
-export function getOwnerDocumentFromRootContainer(
-  rootContainerElement: Element | Document | DocumentFragment,
-): Document {
-  return rootContainerElement.nodeType === DOCUMENT_NODE
-    ? (rootContainerElement: any)
-    : rootContainerElement.ownerDocument;
 }
 
 function noop() {}
@@ -449,136 +421,6 @@ function setProp(
       }
     }
   }
-}
-
-// creates a script element that won't execute
-export function createPotentiallyInlineScriptElement(
-  ownerDocument: Document,
-): Element {
-  // Create the script via .innerHTML so its "parser-inserted" flag is
-  // set to true and it does not execute
-  const div = ownerDocument.createElement('div');
-  if (__DEV__) {
-    if (enableTrustedTypesIntegration && !didWarnScriptTags) {
-      console.error(
-        'Encountered a script tag while rendering React component. ' +
-          'Scripts inside React components are never executed when rendering ' +
-          'on the client. Consider using template tag instead ' +
-          '(https://developer.mozilla.org/en-US/docs/Web/HTML/Element/template).',
-      );
-      didWarnScriptTags = true;
-    }
-  }
-  div.innerHTML = '<script><' + '/script>'; // eslint-disable-line
-  // This is guaranteed to yield a script element.
-  const firstChild = ((div.firstChild: any): HTMLScriptElement);
-  const element = div.removeChild(firstChild);
-  return element;
-}
-
-export function createSelectElement(
-  props: Object,
-  ownerDocument: Document,
-): Element {
-  let element;
-  if (typeof props.is === 'string') {
-    element = ownerDocument.createElement('select', {is: props.is});
-  } else {
-    // Separate else branch instead of using `props.is || undefined` above because of a Firefox bug.
-    // See discussion in https://github.com/facebook/react/pull/6896
-    // and discussion in https://bugzilla.mozilla.org/show_bug.cgi?id=1276240
-    element = ownerDocument.createElement('select');
-  }
-  if (props.multiple) {
-    element.multiple = true;
-  } else if (props.size) {
-    // Setting a size greater than 1 causes a select to behave like `multiple=true`, where
-    // it is possible that no option is selected.
-    //
-    // This is only necessary when a select in "single selection mode".
-    element.size = props.size;
-  }
-  return element;
-}
-
-// Creates elements in the HTML namesapce
-export function createHTMLElement(
-  type: string,
-  props: Object,
-  ownerDocument: Document,
-): Element {
-  if (__DEV__) {
-    switch (type) {
-      case 'script':
-      case 'select':
-        console.error(
-          'createHTMLElement was called with a "%s" type. This type has special creation logic in React and should use the create function implemented specifically for it. This is a bug in React.',
-          type,
-        );
-        break;
-      case 'svg':
-      case 'math':
-        console.error(
-          'createHTMLElement was called with a "%s" type. This type must be created with Document.createElementNS which this method does not implement. This is a bug in React.',
-          type,
-        );
-    }
-  }
-
-  let isCustomComponentTag;
-
-  let element: Element;
-  if (__DEV__) {
-    isCustomComponentTag = isCustomComponent(type, props);
-    // Should this check be gated by parent namespace? Not sure we want to
-    // allow <SVG> or <mATH>.
-    if (!isCustomComponentTag && type !== type.toLowerCase()) {
-      console.error(
-        '<%s /> is using incorrect casing. ' +
-          'Use PascalCase for React components, ' +
-          'or lowercase for HTML elements.',
-        type,
-      );
-    }
-  }
-
-  if (typeof props.is === 'string') {
-    element = ownerDocument.createElement(type, {is: props.is});
-  } else {
-    // Separate else branch instead of using `props.is || undefined` above because of a Firefox bug.
-    // See discussion in https://github.com/facebook/react/pull/6896
-    // and discussion in https://bugzilla.mozilla.org/show_bug.cgi?id=1276240
-    element = ownerDocument.createElement(type);
-  }
-
-  if (__DEV__) {
-    if (
-      !isCustomComponentTag &&
-      // $FlowFixMe[method-unbinding]
-      Object.prototype.toString.call(element) ===
-        '[object HTMLUnknownElement]' &&
-      !hasOwnProperty.call(warnedUnknownTags, type)
-    ) {
-      warnedUnknownTags[type] = true;
-      console.error(
-        'The tag <%s> is unrecognized in this browser. ' +
-          'If you meant to render a React component, start its name with ' +
-          'an uppercase letter.',
-        type,
-      );
-    }
-  }
-
-  return element;
-}
-
-export function createTextNode(
-  text: string,
-  rootContainerElement: Element | Document | DocumentFragment,
-): Text {
-  return getOwnerDocumentFromRootContainer(rootContainerElement).createTextNode(
-    text,
-  );
 }
 
 export function setInitialProperties(
