+++ packages/react-dom-bindings/src/client/ReactDOMHostConfig.js
@@ -44,10 +44,6 @@ import {
 export {detachDeletedInstance};
 import {hasRole} from './DOMAccessibilityRoles';
 import {
-  createHTMLElement,
-  createPotentiallyInlineScriptElement,
-  createSelectElement,
-  createTextNode,
   setInitialProperties,
   diffProperties,
   updateProperties,
@@ -59,7 +55,6 @@ import {
   warnForDeletedHydratableText,
   warnForInsertedHydratedElement,
   warnForInsertedHydratedText,
-  getOwnerDocumentFromRootContainer,
 } from './ReactDOMComponent';
 import {getSelectionInformation, restoreSelection} from './ReactInputSelection';
 import setTextContent from './setTextContent';
@@ -90,6 +85,7 @@ import {
   enableScopeAPI,
   enableFloat,
   enableHostSingletons,
+  enableTrustedTypesIntegration,
 } from 'shared/ReactFeatureFlags';
 import {
   HostComponent,
@@ -105,6 +101,7 @@ import {
   getValueDescriptorExpectingObjectForWarning,
   getValueDescriptorExpectingEnumForWarning,
 } from '../shared/ReactDOMResourceValidation';
+import isCustomComponent from '../shared/isCustomComponent';
 
 export type Type = string;
 export type Props = {
@@ -119,6 +116,9 @@ export type Props = {
   left?: null | number,
   right?: null | number,
   top?: null | number,
+  is?: string,
+  size?: number,
+  multiple?: boolean,
   ...
 };
 type RawProps = {
@@ -181,6 +181,14 @@ let eventsEnabled: ?boolean = null;
 let selectionInformation: null | SelectionInformation = null;
 
 export * from 'react-reconciler/src/ReactFiberHostConfigWithNoPersistence';
+
+function getOwnerDocumentFromRootContainer(
+  rootContainerElement: Element | Document | DocumentFragment,
+): Document {
+  return rootContainerElement.nodeType === DOCUMENT_NODE
+    ? (rootContainerElement: any)
+    : rootContainerElement.ownerDocument;
+}
 
 export function getRootHostContext(
   rootContainerInstance: Container,
@@ -285,13 +293,28 @@ export function createHoistableInstance(
   const ownerDocument = getOwnerDocumentFromRootContainer(
     rootContainerInstance,
   );
-  const domElement: Instance = createHTMLElement(type, props, ownerDocument);
+
+  const domElement: Instance = ownerDocument.createElement(type);
   precacheFiberNode(internalInstanceHandle, domElement);
   updateFiberProps(domElement, props);
   setInitialProperties(domElement, type, props);
   markNodeAsHoistable(domElement);
   return domElement;
 }
+
+let didWarnScriptTags = false;
+const warnedUnknownTags: {
+  [key: string]: boolean,
+} = {
+  // There are working polyfills for <dialog>. Let people use it.
+  dialog: true,
+  // Electron ships a custom <webview> tag to display external web content in
+  // an isolated frame and process.
+  // This tag is not present in non Electron environments such as JSDom which
+  // is often used for testing purposes.
+  // @see https://electronjs.org/docs/api/webview-tag
+  webview: true,
+};
 
 export function createInstance(
   type: string,
@@ -334,20 +357,99 @@ export function createInstance(
       break;
     default:
       switch (type) {
-        case 'svg':
+        case 'svg': {
           domElement = ownerDocument.createElementNS(SVG_NAMESPACE, type);
           break;
-        case 'math':
+        }
+        case 'math': {
           domElement = ownerDocument.createElementNS(MATH_NAMESPACE, type);
           break;
-        case 'script':
-          domElement = createPotentiallyInlineScriptElement(ownerDocument);
+        }
+        case 'script': {
+          // Create the script via .innerHTML so its "parser-inserted" flag is
+          // set to true and it does not execute
+          const div = ownerDocument.createElement('div');
+          if (__DEV__) {
+            if (enableTrustedTypesIntegration && !didWarnScriptTags) {
+              console.error(
+                'Encountered a script tag while rendering React component. ' +
+                  'Scripts inside React components are never executed when rendering ' +
+                  'on the client. Consider using template tag instead ' +
+                  '(https://developer.mozilla.org/en-US/docs/Web/HTML/Element/template).',
+              );
+              didWarnScriptTags = true;
+            }
+          }
+          div.innerHTML = '<script><' + '/script>'; // eslint-disable-line
+          // This is guaranteed to yield a script element.
+          const firstChild = ((div.firstChild: any): HTMLScriptElement);
+          domElement = div.removeChild(firstChild);
           break;
-        case 'select':
-          domElement = createSelectElement(props, ownerDocument);
+        }
+        case 'select': {
+          if (typeof props.is === 'string') {
+            domElement = ownerDocument.createElement('select', {is: props.is});
+          } else {
+            // Separate else branch instead of using `props.is || undefined` above because of a Firefox bug.
+            // See discussion in https://github.com/facebook/react/pull/6896
+            // and discussion in https://bugzilla.mozilla.org/show_bug.cgi?id=1276240
+            domElement = ownerDocument.createElement('select');
+          }
+          if (props.multiple) {
+            domElement.multiple = true;
+          } else if (props.size) {
+            // Setting a size greater than 1 causes a select to behave like `multiple=true`, where
+            // it is possible that no option is selected.
+            //
+            // This is only necessary when a select in "single selection mode".
+            domElement.size = props.size;
+          }
           break;
-        default:
-          domElement = createHTMLElement(type, props, ownerDocument);
+        }
+        default: {
+          let isCustomComponentTag;
+
+          if (__DEV__) {
+            isCustomComponentTag = isCustomComponent(type, props);
+            // Should this check be gated by parent namespace? Not sure we want to
+            // allow <SVG> or <mATH>.
+            if (!isCustomComponentTag && type !== type.toLowerCase()) {
+              console.error(
+                '<%s /> is using incorrect casing. ' +
+                  'Use PascalCase for React components, ' +
+                  'or lowercase for HTML elements.',
+                type,
+              );
+            }
+          }
+
+          if (typeof props.is === 'string') {
+            domElement = ownerDocument.createElement(type, {is: props.is});
+          } else {
+            // Separate else branch instead of using `props.is || undefined` above because of a Firefox bug.
+            // See discussion in https://github.com/facebook/react/pull/6896
+            // and discussion in https://bugzilla.mozilla.org/show_bug.cgi?id=1276240
+            domElement = ownerDocument.createElement(type);
+          }
+
+          if (__DEV__) {
+            if (
+              !isCustomComponentTag &&
+              // $FlowFixMe[method-unbinding]
+              Object.prototype.toString.call(domElement) ===
+                '[object HTMLUnknownElement]' &&
+              !hasOwnProperty.call(warnedUnknownTags, type)
+            ) {
+              warnedUnknownTags[type] = true;
+              console.error(
+                'The tag <%s> is unrecognized in this browser. ' +
+                  'If you meant to render a React component, start its name with ' +
+                  'an uppercase letter.',
+                type,
+              );
+            }
+          }
+        }
       }
   }
   precacheFiberNode(internalInstanceHandle, domElement);
@@ -429,7 +531,9 @@ export function createTextInstance(
     const hostContextDev = ((hostContext: any): HostContextDev);
     validateDOMNesting(null, text, hostContextDev.ancestorInfo);
   }
-  const textNode: TextInstance = createTextNode(text, rootContainerInstance);
+  const textNode: TextInstance = getOwnerDocumentFromRootContainer(
+    rootContainerInstance,
+  ).createTextNode(text);
   precacheFiberNode(internalInstanceHandle, textNode);
   return textNode;
 }
